test: pet tests should work, but somehow they don't, even though they should
