+++ tests/unit/pets/pets.controller.unit-spec.ts
@@ -1,10 +1,9 @@
 import { Test } from "@nestjs/testing";
-import { CreatePetDto } from "src/modules/pets/dto";
-import { PetsController } from "src/modules/pets/pets.controller";
-import { PetsService } from "src/modules/pets/pets.service";
-import { petStub } from "src/modules/pets/stubs/";
-
-jest.mock("../../../src/modules/pets/pets.service.ts");
+import { CreatePetDto } from "../../../src/modules/pets/dto";
+import { PetsController } from "../../../src/modules/pets/pets.controller";
+import { PetsService } from "../../../src/modules/pets/pets.service";
+import { petStub } from "../../../src/modules/pets/stubs/";
+import { PetsServiceMock } from "../../../src/modules/pets/__mocks__";
 
 describe("PetsController Unit", () => {
   let petsController: PetsController;
@@ -13,7 +12,12 @@ describe("PetsController Unit", () => {
   beforeAll(async () => {
     const moduleRef = await Test.createTestingModule({
       controllers: [PetsController],
-      providers: [PetsService],
+      providers: [
+        {
+          provide: PetsService,
+          useClass: PetsServiceMock,
+        },
+      ],
     }).compile();
 
     petsController = moduleRef.get<PetsController>(PetsController);
@@ -28,17 +32,17 @@ describe("PetsController Unit", () => {
     it("when create is called, it should call PetsService", async () => {
       const createPetDto: CreatePetDto = {
         name: "test",
-        speciesId: "0334f499-a655-4d62-8f0d-2b68a5a1c93e",
+        speciesId: "testSpeciesID",
         age: 12,
         gender: "m",
-        breed: "akita",
+        breed: "testBreed",
         isCastrated: false,
-        locationId: "00395870-a458-4fba-8bbf-e91b840524ab",
+        locationId: "testCityID",
         description: "test",
         petSize: "MEDIUM",
       };
-
       const pet = petsController.create(createPetDto);
+
       expect(petsService.create).toHaveBeenCalledWith(createPetDto);
       expect(pet).toEqual(petStub());
     });
